Ownership of tiles and color based on ownership
Added ownership to every tile, randomly assigned on startup (this could
be our bonus thing to get an A on the demo if we have the starting zones
random rather than fixed). Tiles are now owned by a player, and the tint
of their material is determined by that player's color. We only need to
create textures representing each land type (I tried this but the
textures seem to be wiped from the prefabs on load, I'll look into it
tomorrow) or we can add another mesh to each prefab that represents the
tile type. I also moved some files into other folders.