updated Village.cs and Tile.cs , added Structure.cs (enum)
Added new enum Structure:
The tile that is the hub of a village is of this type. I have added
Tower and we can add more as we see fit. In addition the type
NoStructure also exists (default in initialization)
Village.cs:  tombPhase() and delete() updated
Tile.cs: added a bunch of methods from class diagram