Combat
Units are now unable to move to a tile where these is another unit
already present.
Units are also unable to move to a tile that has a STRONGER enemy in a 1
tile radius.
If the enemy is weaker, the unit moves and kills the enemy. Right now it
just destroys the enemy gameobject, still need to replace with tombstone
and decrement upkeep of the village.